+++ docs/changelog.md
@@ -79,7 +79,7 @@
 - make `ansi` module public.
 
 ### text::layout
-- new types: `TextBreakKind`, `TextBreakMode`, `TextLine`, `TextLineIter`, `TextElideMode`, `TextSegment`, `TextSegmentKind`, `TextSymbolConfig`, `TextWrapIter`, `TextWidthMode`.
+- new types: `TextBreakKind`, `TextBreakMode`, `TextLine`, `TextLineIter`, `TextElideMode`, `TextSegment`, `TextSegmentKind`, `TextSymbolConfig`, `TextWrapIter`, `TextCellWidthMode`.
 
 ### text::str
 - update `Str`: